Rename JSONParser => JSONSingleContainerParser
close DCOS-14527
