+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
